requirements.txt
fastapi
uvicorn
streamlit
chromadb
sentence-transformers
groq
python-dotenv
faker
numpy
pandas
pydantic
requests
plotly